Fill in the correct d_off in rtems_rfs_dir_read.
rtems_rfs_dir_read searches the directory inode's entries list starting at the
specified offset until an empty entry (last entry) is encountered. It fills in
a struct dirent with the name of the entry, length of the name, ino of the
entry, and the absolute offset of the entry in the parent directory's entries
list.

Unfortunately, the stock implementation of rtems_rfs_dir_read returns a
somewhat arbitrary offset (as dirent::d_off), while rtems_rfs_dir_lookup_ino
always returns the correct offset.

This change fixes that logic so the returned offset is accurate.

Tested by comparing the offset returned in dirent with the result of
rtems_rfs_dir_lookup_ino.